remove neon instant from pr
