+++ docs/2.connectors/1.index.md
@@ -12,7 +12,6 @@ Currently supported connectors:
 - [Cloudflare D1](/connectors/cloudflare)
 - [LibSQL](/connectors/libsql)
 - [Neon](/connectors/neon)
-- [Neon Instant](/connectors/neon-instant)
 - [PlanetScale](/connectors/planetscale)
 - [PostgreSQL](/connectors/postgresql)
 - [Prisma Postgres](/connectors/postgresql#prisma-postgres)
@@ -37,7 +36,6 @@ DB0 itself has no runtime dependencies. Connectors that talk to a third-party da
 | [`libsql`](/connectors/libsql)                                                     | [`@libsql/client`](https://www.npmjs.com/package/@libsql/client)               |
 | [`mysql2`](/connectors/mysql)                                                      | [`mysql2`](https://www.npmjs.com/package/mysql2)                               |
 | [`neon`](/connectors/neon)                                                         | [`@neondatabase/serverless`](https://www.npmjs.com/package/@neondatabase/serverless) |
-| [`neon-instant`](/connectors/neon-instant)                                         | [`@neondatabase/serverless`](https://www.npmjs.com/package/@neondatabase/serverless), [`neon-new`](https://www.npmjs.com/package/neon-new) _(optional)_ |
 | [`node-sqlite`](/connectors/sqlite#node-sqlite)                                    | — (built into Node.js and Deno)                                                |
 | [`pglite`](/connectors/pglite)                                                     | [`@electric-sql/pglite`](https://www.npmjs.com/package/@electric-sql/pglite)   |
 | [`planetscale`](/connectors/planetscale)                                           | [`@planetscale/database`](https://www.npmjs.com/package/@planetscale/database) |
