+++ package.json
@@ -63,7 +63,6 @@
     "kysely": "^0.29.5",
     "mlly": "^1.8.2",
     "mysql2": "^3.23.3",
-    "neon-new": "^0.15.0",
     "obuild": "^0.4.38",
     "pathe": "^2.0.3",
     "pg": "^8.23.0",
